https://github.com/ImageMagick/ImageMagick/pull/3177

+++ MagickCore/thread-private.h
@@ -61,9 +61,9 @@ static inline int GetMagickNumberThreads(const Image *source,
        (GetImagePixelCacheType(source) != MapCache)) || 
       ((GetImagePixelCacheType(destination) != MemoryCache) && 
        (GetImagePixelCacheType(destination) != MapCache)))
-    return(MagickMax(MagickMin(GetMagickResourceLimit(ThreadResource),2),1));
-  return(MagickMax(MagickMin((ssize_t) GetMagickResourceLimit(ThreadResource),
-    (ssize_t) (chunk)/64),1));
+    return((int) MagickMax(MagickMin(GetMagickResourceLimit(ThreadResource),2),1));
+  return((int) MagickMax(MagickMin((ssize_t) GetMagickResourceLimit(
+    ThreadResource),(ssize_t) (chunk)/64),1));
 }
 
 static inline MagickThreadType GetMagickThreadId(void)
